fail test for example 1 and 2
